chore: move `JsonConverter`
